+++ example/lib/main.dart
@@ -182,6 +182,8 @@ class _MailPageState extends State<MailPage> {
                                       key: _one,
                                       description: 'Tap to see menu options',
                                       disableDefaultTargetGestures: true,
+                                      onBarrierClick: () =>
+                                          debugPrint('Barrier clicked'),
                                       child: GestureDetector(
                                         onTap: () =>
                                             debugPrint('menu button clicked'),