+++ lib/src/showcase.dart
@@ -238,6 +238,12 @@ class Showcase extends StatefulWidget {
   /// Provides text direction of tooltip description.
   final TextDirection? descriptionTextDirection;
 
+  /// Provides a callback when barrier has been clicked.
+  ///
+  /// Note-: Even if barrier interactions are disabled, this handler
+  /// will still provide a callback.
+  final VoidCallback? onBarrierClick;
+
   const Showcase({
     required this.key,
     required this.description,
@@ -281,6 +287,7 @@ class Showcase extends StatefulWidget {
     this.descriptionPadding,
     this.titleTextDirection,
     this.descriptionTextDirection,
+    this.onBarrierClick,
   })  : height = null,
         width = null,
         container = null,
@@ -317,6 +324,7 @@ class Showcase extends StatefulWidget {
     this.onTargetDoubleTap,
     this.disableDefaultTargetGestures = false,
     this.tooltipPosition,
+    this.onBarrierClick,
   })  : showArrow = false,
         onToolTipClick = null,
         scaleAnimationDuration = const Duration(milliseconds: 300),
@@ -486,6 +494,7 @@ class _ShowcaseState extends State<Showcase> {
             if (!showCaseWidgetState.disableBarrierInteraction) {
               _nextIfAny();
             }
+            widget.onBarrierClick?.call();
           },
           child: ClipPath(
             clipper: RRectClipper(